[TW010] - Core/Arena: Anonymous arena queue
This makes sure that when someone joins the arena queue, the world announcement does not mention which teams joined. The exit announcement is completely removed. Also it is not possible to discover who is in arena through the guild, friend or who list because it will show that they are in Dalaran.